Fix parameter validation logic, rename AllowedValue value
Before any parameter value having allowed values and no minLength or
maxLength set would pass validation since if the allowed values check
did not pass it would then fall through to the length and regex, whose
default was @any regex.  Now, if allowed values are provided, the
provided value is solely checked against them.  If they are not
provided, the other checks kick in.  Fix tests that relied on the bad
behavior.

The value property of the AllowedValue was called `name`, it is now
`value`.

Add types to public methods.