Deserialize linked json to arbitrary depth
Previously we were creating a new object to update the link. This means
that subsequent updates to those objects were not being reflected by
objects that referenced them: as they were no longer references. This
introduces a wrapper that allows us to override the link name whilst
still holding a reference to the original object so that updates are
reflected across the tree.