Make name optional on a regex
This biggest reason for this is to give name and regex a different type so I don't get them mixed up
But also, it's cromulent to define a regex of things that you don't want to name.

I made another constructor in the Object so that I don't have to change all the usages right now.