Raise an InvalidRugTestScenarioName when scenario name is empty
This should address #71 when the scenario name is not set.
By raising that specific error, the user gets a message that
the name is invalid (since an empty name is considered as such).
This was much trickier than expected because scala Combinators regex
used before wasn't stopping at the end of the scenario line but
would swallow everythong down to the next production. David and I
figured we'd need to be a little more explicit about the actual
parsing rule.