Add ArchiveRoot step, TypeDoc test
Add a well-known step for the DSL ArchiveRoot.

Include the test directory in the generated TypeDoc (is there a reason
operations and tree are not?).

Clean up TypeScript formatting.

Make TypeDoc on ScenarioWorld.modificationsMade clear that the initial
population of the project by a generator does not count as a
modification.

Possible closes #426