Parameters are not optional, fix regression
Calling `editWith()` and `generateWith()` in TypeScript without a
second argument resulted in an exception since the underlying method
in Scala did not provide a default value.  Given @alankstwewart
finding issued with method overloading, to fix I just made the second
argument not optional in the TypeScript interface.

When EditorWithParametersSteps got moved to its own file, it lost its
parameter and started calling the wrong editor.