Allow testing of aborted scenarios (#571)
Add special case of Gherkin `Then` case allowing tests to see if a
scenario was aborted due to an exception being thrown.  Add TypeScript
`Then the scenario was aborted` step implementation so it can be
discovered but it is not actually used.

Add Typescript `Given nothing` step implementation for command handler
testing.

Fix test that was passing due to the wrong failure, adding assertion
to make sure it does not happen again.

Clean up change log.