TypeUnderFile actually returns an OTTN not a MCTV
I had to mark a test PendingUntilFixed for this change
I know the change is good; the test (ReflectiveTypeOperationFinderTest)
however isn't testing what it wanted to.

This should work afaict but i don't know
what is going on. I changed the runtimeClass in TypeUnderFile,
that is what affects it. But I dunno why all the operations show up now
in both allOperations and operations